added feature to print and validate local and remote sha1 digests
By definining the new subinfo class property targetDigests local sha1 digests could be added to targets. 
The property could hold a string or a list.
The new subinfo class property targetDigestUrls supports remote sha1 digest files for targets. 
The property could hold a string or a list. 

The ArchiveSource class reads digests remote or local and validates the relating files before unpacking.
If no local or remote digest for a file is available, unpack() prints the digest of the related files in 
a way that the digest could be easily added to the portage script. 

svn path=/trunk/kdesupport/emerge/; revision=1114393