utils / manifests: Cleanup, fixes
It happened to me that a manifest file did not exist when emerge tried to
open it and emerge died, so I looked closer at this code.

After some refactoring, this is fixed. Also:

- removing code duplication: less source code
- more complete docstrings
- if a manifest file appears twice in *.mft, emerge tried to delete it
  twice, so it logged a warning about a missing file. Fixed.
- in the same situation, if one of the entries had a hash and the other
  not, the hash was not checked before deleting. Now it is.

those 2 problems happen with dev-util/uactools, unmerging this will now result in a
warning that uactools-20101219-bin.mft has a wrong hash (which it has).

This uncovers another problem about uactools:
this hash should not exist in the manifest file at all - if file X contains
its own hash, it is always wrong (well, in practice at least)

svn path=/trunk/kdesupport/emerge/; revision=1223382