Add a portable function for TzSpecificLocalTimeToSystemTime in kjs date object.
This makes it possible to compile kdelibs again without a hacked mingw-w32 and with mingw.org
Yes that introduces a dependency to QT in KJS so it's not really upstreamable. But since I do not understand why it is neccessary there in the first place I am not going to pursue this issue.

svn path=/trunk/kdesupport/emerge/; revision=1234175