added support for "all" package filename version.
Generated binary package versions are determined from the package filename in case they are not overridden by the package itself or the --options command line option or when using ...HEAD target.

To not been forced to change the package filename on every package version change, the "all" filename version could be used.

If a package filename contains the "all" version, the binary package filename version is determined from the current build target.