qt5: Use 5.6 branch again
TQtC decided to mess up the 5.6.1 release and we now end up with an hot
fix release called 5.6.1-1. While the 5.6.1-1 tarballs got renamed, the contents
of the tarballs still have the 5.6.1 in folder names.

Emerge cannot cope with that. And I don't think it's worth fixing the
recipe just for this release => Let's just default to 5.6 branch for now, and wait
for 5.6.2.