Delay target evaluation
Delay the taget evaluation and no longer perform it in the constructor.
This fixes the `You defined an invalid target` error when resolving the dependencies.
The error will only occure when the actual package is accessed.
The motivation for this change is restoring a craft shelf that is no longer available in the current version of craft
but in the version that will be restored from the shelf.