Add fallback value to XDG_DATA_DIRS
On some distributions (e.g., Debian 11), $XDG_DATA_DIRS is not defined by
default. This is fine as long as the path is not supposed to be prepended
with a custom value. However, if it is not defined, the variable then
receives an actual value. This causes most UI frameworks (including Gtk+ and
Qt) to only use the directories from this path and no longer use a default
value defined internally.

An easy fix is to just always append the default value defined within the
XDG base directories specification. That way, even if the value is not
defined by default, it will fall back to /usr/local/share:/usr/share.

This commit fixes opening browsers with Qt's QDesktopServices from within
an AppImage. Before, no browser was opened and no error was reported to the
application (an error message was typically written to the console, though).