Don't make Craft fail on bad-encoded curl output
As experienced with mpfr in CI mode.

Proofed by @kloecker:
You can get the same undecodable character when running `curl --head -v [...]`. So this looks like a bug in curl. curl prints a single byte for Î, i.e. it definitely doesn't use utf-8. Firefox shows the character correctly and openssl print the character as \C3\8Ele-de-France, so it is almost certain that the certificate also uses UTF-8.
In other words: It's not our fault and we cannot do anything about it and need to work it around.