Set up Meson for cross compilation
Previously Meson was not set up for cross compilation, so libraries
that use Meson (like fribidi) breaks Android builds due to them
compiling to the host architecture.

Meson uses a text file, similar to CMake, for selecting toolchain
options for cross-compilation. Unlike CMake, the toolchain files
are completely static so we generate one in /etc with the
necessary constants set.