Use ECM Android toolchain from target installation when available
This allows updates to the toolchain without requiring rebuilds of the
host image and matches what the CI does. The toolchain file from the
host image is still used if ECM hasn't been built yet, but since ECM
doesn't need much from in there this usually is more than enough for
bootstrapping.