Allowing the MacPackager to handle universal binaries for KStars
Several of our 3rdParty Libraries for drivers for KStars are provided by manufacturers as universal binaries so that they work on both Intel and arm processors.  Since they are provided as binaries, craft can't build them with just one architecture.  When packaging, currently craft has an assertion that sometimes happens on universal binaries.  This should fix that and allow it to proceed.